Use epoch-prefixed i64 message ids shared per session
Message ids are now i64 values built as (epoch << 32) | counter,
replacing the i32 MessageID. Epoch is a per-session run count; the
counter is a per-run atomic shared by the root conversation and every
subagent, so ids are globally unique across the whole session dir
without any on-disk scan.

A single UniqueIdGenerator per session is created by
ConversationManager::ensure_id_generator from the root session dir.
The epoch-file protocol (read msg-id-epoch, bump, tmp + fsync +
rename + dir fsync) runs once at session start, before any id is
minted, keeping the ordering invariant: no event with the new epoch
is written before the epoch file durably reads it. Legacy sessions
migrate for free (old ids < 2^32; first resume bumps to epoch 1).

Also:
- Remove ConversationState.msg_id_counter; resume no longer seeds a
  high-water counter. close_stale_running_items no longer scans
  tool-call files or derives max ids; synthetic stale-close events
  mint from the session generator.
- Add a fast display.jsonl pre-filter (tcode-runtime display_scan)
  used by the stale-close scan and the tree pane, skipping the ~97%
  of lines that are chunk/thinking/arg-chunk events.
- Speed up the nvim display pane's initial render by deferring
  per-chunk buffer writes to collapse points.
- Branching copies msg-id-epoch so the branch resumes one epoch
  above every retained id; validate_branch drops the id-order and
  state-counter checks; the branch msg-id CLI argument is now i64.
- Tests: epoch/legacy/crash/synthetic-id coverage, RAII TestDir
  cleanup, display_scan marker tests.

+++ tcode/src/main.rs
@@ -166,7 +166,7 @@ enum Commands {
     /// Branch off a new independent session at a user message (clone history up to, not including, that message)
     Branch {
         /// The display msg_id of the target user message
-        msg_id: i32,
+        msg_id: i64,
     },
     /// Cancel a running tool call
     CancelTool {
